update Vue to 2.1.3 and use build-time template compilation
