--- vue-v2.0.7/index.html
@@ -1,58 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Vue.js v2.0.7</title>
-  <link href="../css/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet"/>
-  <link href="../css/main.css" rel="stylesheet"/>
-</head>
-<body>
-  <div id='main'>
-      <div class="container">
-          <div class="jumbotron">
-              <div class="row">
-                  <div class="col-md-6">
-                      <h1>Vue.js v2.0.7</h1>
-                  </div>
-                  <div class="col-md-6">
-                      <div class="row">
-                          <div class="col-sm-6 smallpad">
-                            <button type="button" class="btn btn-primary btn-block" id="run" v-on:click="run">Create 1,000 rows</button>
-                          </div>
-                          <div class="col-sm-6 smallpad">
-                              <button type="button" class="btn btn-primary btn-block" id="runlots" v-on:click="runLots">Create 10,000 rows</button>
-                          </div>
-                          <div class="col-sm-6 smallpad">
-                              <button type="button" class="btn btn-primary btn-block" id="add" v-on:click="add">Append 1,000 rows</button>
-                          </div>
-                          <div class="col-sm-6 smallpad">
-                              <button type="button" class="btn btn-primary btn-block" id="update" v-on:click="update">Update every 10th row</button>
-                          </div>
-                          <div class="col-sm-6 smallpad">
-                              <button type="button" class="btn btn-primary btn-block" id="clear" v-on:click="clear">Clear</button>
-                          </div>
-                          <div class="col-sm-6 smallpad">
-                              <button type="button" class="btn btn-primary btn-block" id="swaprows" v-on:click="swapRows">Swap Rows</button>
-                          </div>
-                      </div>
-                  </div>
-              </div>
-          </div>
-          <table class="table table-hover table-striped test-data">
-              <tbody>
-                  <tr v-for="item in rows" v-bind:class="{'danger' : item.id === selected}">
-                      <td class="col-md-1">{{item.id}}</td>
-                      <td class="col-md-4">
-                          <a v-on:click="select(item)">{{item.label}}</a>
-                      </td>
-                      <td class="col-md-1"><a v-on:click="remove(item)"><span class="glyphicon glyphicon-remove" aria-hidden="true"></span></a></td>
-                      <td class="col-md-6"></td>
-                  </tr>
-              </tbody>
-          </table>
-          <span class="preloadicon glyphicon glyphicon-remove" aria-hidden="true"></span>
-      </div>
-  </div>
-  <script src='dist/main.js'></script>
-</body>
-</html>
