+++ vue-v2.1.3/package.json
@@ -21,9 +21,12 @@
     "babel-loader": "6.2.8",
     "babel-preset-es2015": "6.18.0",
     "babel-preset-stage-0": "6.16.0",
-    "webpack": "1.13.3"
+    "webpack": "1.13.3",
+    "vue-loader": "10.0.2",
+    "vue-template-compiler": "2.1.3",
+    "css-loader": "0.26.0"
   },
   "dependencies": {
-    "vue": "2.0.7"
+    "vue": "2.1.3"
   }
 }
