+++ vue-v2.1.3/src/main.es6.js
@@ -1,0 +1,7 @@
+import Vue from 'vue'
+import App from './App.vue'
+
+new Vue({
+    el: '#main',
+    render: h => h(App)
+})
