+++ vue-v2.1.3/src/store.es6.js
